[onert] Change ModelIndex value type (#9794)
This commit changes ModelIndex value type from uint32_t to uint16_t.

This commit includes updating hash function for ModelIndex-SubgraphIndex pair.

ONE-DCO-1.0-Signed-off-by: Hyeongseok Oh <hseok82.oh@samsung.com>

+++ runtime/onert/core/include/ir/Index.h
@@ -39,7 +39,7 @@ struct SubgraphIndexTag;
 using SubgraphIndex = ::onert::util::Index<uint32_t, SubgraphIndexTag>;
 
 struct ModelIndexTag;
-using ModelIndex = ::onert::util::Index<uint32_t, ModelIndexTag>;
+using ModelIndex = ::onert::util::Index<uint16_t, ModelIndexTag>;
 
 template <typename IndexType>
 std::ostream &_index_print_impl(std::ostream &o, const std::string &prefix, IndexType index)
